// packages/wrangler-shared/src/config/environment.d.ts
export type ContainerApp = {
    name: string;
    instances: number;
    scheduling_policy?: "regional" | "moon";
    configuration: {
        image: string;
        labels?: {
            name: string;
            value: string;
        }[];
        secrets?: {
            name: string;
            type: "env";
            secret: string;
        }[];
    };
    constraints?: {
        regions?: string[];
        cities?: string[];
        tier?: number;
    };
};